requirements.txt
# ===========================================
# NUVARU PLATFORM - PRODUCTION REQUIREMENTS
# ===========================================

# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.0.350
langchain-community==0.0.1
langchain-core==0.1.0
numpy==1.24.3
scikit-learn==1.3.2

# LLM and AI
openai==1.3.7
tiktoken==0.5.2

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8
bcrypt==4.1.2

# HTTP client and utilities
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0

# AWS S3 support
boto3==1.34.0
botocore==1.34.0

# PDF processing
PyPDF2==3.0.1
PyCryptodome==3.19.0

# Monitoring and logging
structlog==23.2.0

# Production server
gunicorn==21.2.0
